agenttel-cli/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agenttel-cli"
version = "0.1.0"
description = "CLI tool for analyzing Java codebases and generating AgentTel configuration"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.25",
]

[project.scripts]
agenttel = "agenttel_cli.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
